requirements.txt
google-genai>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
tavily-python>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0